// generated/src/app/api-models/stage-psclm-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePsclmDtl {
  batchId: string;
  transactionId: number;
  lineNo: number;
  subLineCode: string;
  lineStatus: string;
  detailSvcDate: Date;
  svcThruDate: Date;
  placeOfService: string;
  diagnosisCodePtr1: number;
  diagnosisCodePtr2: number;
  diagnosisCodePtr3: number;
  diagnosisCodePtr4: number;
  procedureCode: string;
  procedureModifier1: string;
  procedureModifier2: string;
  procedureModifier3: string;
  procedureModifier4: string;
  quantity: number;
  billedAmt: number;
  allowedFactor: number;
  allowedAmt: number;
  allowedRsn: string;
  notCoveredAmt: number;
  notCoveredRsn: string;
  copayAmt: number;
  copayRsn: string;
  copay2Amt: number;
  copay2Rsn: string;
  deductibleAmt: number;
  deductibleRsn: string;
  otherCarrierAmt: number;
  otherCarrierRsn: string;
  withholdAmt: number;
  netAmt: number;
  primaryPaidAmt: number;
  ocAllowedAmt: number;
  adjustmentRsn: string;
  checkDate: Date;
  anesStartTime: number;
  anesEndTime: number;
  anesTotalTime: number;
  typeOfService: string;
  epsdt: string;
  familyPlanningInd: string;
  emergencySvcsInd: string;
  cobInd: string;
  adjudDate: Date;
  reservedLocalUseDtl: string;
  upnNo: string;
  emgCode: string;
  interestAmt: number;
  discountAmt: number;
  paidNetAmt: number;
  uncleanDays: number;
  overrideUncleanDays: number;
  claimStatusOut: string;
  processingStatusOut: string;
  procCodeQualifier: string;
  dmeProcCodeQualifier: string;
  dmeProcedureCode: string;
  dmeUnitMeasure: string;
  dmeQuantity: number;
  dmeRentalPrice: number;
  dmePurchasePrice: number;
  dmeRentalFrequency: string;
  dateLatestVisit: Date;
  latestHemoTestDate: Date;
  latestCreatineTestDate: Date;
  mammographyCertNo: string;
  purchasedSvcChargeProvId: string;
  purchasedSvcChargeAmt: number;
  spinalManipInitTreatmentDt: Date;
  providerLineItemControlNo: string;
  cobPatLiabCvrgAmt: number;
  dtlClaimNumber: string;
  insertDatetime: Date;
}